# doc/astrolabe-api-openapi-source.yaml
...
openapi: 3.0.0
info:
  title: Rift API
  description: The backend API for the ONC chatbot
  version: v1
servers:
  - url: 'http://localhost:5000'
paths:
...
  /api/files:
    post:
      tags:
        - File
      operationId: File_UploadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                ContentType:
                  type: string
                  nullable: true
                ContentDisposition:
                  type: string
                  nullable: true
                Headers:
                  type: array
                  nullable: true
                  items: {}
                Length:
                  type: integer
                  format: int64
                Name:
                  type: string
                  nullable: true
                FileName:
                  type: string
                  nullable: true
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
    get:
      tags:
        - File
      operationId: File_GetAllFiles
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileEntityDto'
  '/api/files/{id}':
    delete:
      tags:
        - File
      operationId: File_DeleteFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/messages:
    post:
      tags:
        - Message
      operationId: Message_CreateMessage
      requestBody:
        x-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
        required: true
        x-position: 1
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /api/messages/guest:
    post:
      tags:
        - Message
      operationId: Message_CreateGuestMessage
      requestBody:
        x-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
        required: true
        x-position: 1
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  '/api/conversations/{conversationId}/messages':
    get:
      tags:
        - Message
      operationId: Message_GetMessages
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  '/api/messages/{messageId}':
    patch:
      tags:
        - Message
      operationId: Message_UpdateMessageCoordinates
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      requestBody:
        x-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCoordinatesRequest'
        required: true
        x-position: 2
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
    delete:
      tags:
        - Message
      operationId: Message_DeleteMessage
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int32
          x-position: 1
      responses:
        '200':
          description: ''
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    FileEntityDto:
      type: object
      additionalProperties: false
      required:
        - fileName
        - uploadedBy
      properties:
        id:
          type: integer
          format: int32
        fileName:
          type: string
          maxLength: 255
          minLength: 1
        createdAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
          minLength: 1
...
    CreateMessageRequest:
      type: object
      additionalProperties: false
      properties:
        conversationId:
          type: integer
          format: int32
          nullable: true
        content:
          type: string
        xCoordinate:
          type: number
          format: float
        yCoordinate:
          type: number
          format: float
        responseXCoordinate:
          type: number
          format: float
        responseYCoordinate:
          type: number
          format: float
        sourceHandle:
          type: string
        targetHandle:
          type: string
        sessionId:
          type: string
          nullable: true
        sources:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/PartialMessageEdge'
    PartialMessageEdge:
      type: object
      additionalProperties: false
      properties:
        sourceMessageId:
          type: integer
          format: int32
        sourceHandle:
          type: string
        targetHandle:
          type: string
    UpdateCoordinatesRequest:
      type: object
      additionalProperties: false
      properties:
        xCoordinate:
          type: number
          format: float
        yCoordinate:
          type: number
          format: float
...